+++ packages/angular/src/providers/picker-controller.ts
@@ -3,10 +3,13 @@ import { OverlayBaseController } from '@ionic/angular/common';
 import type { PickerOptions } from '@ionic/core';
 import { pickerController } from '@ionic/core';
 
+/**
+ * @deprecated Use the inline ion-picker component instead.
+ */
 @Injectable({
   providedIn: 'root',
 })
-export class PickerController extends OverlayBaseController<PickerOptions, HTMLIonPickerElement> {
+export class PickerController extends OverlayBaseController<PickerOptions, HTMLIonPickerLegacyElement> {
   constructor() {
     super(pickerController);
   }